+++ source/Public/Test-SqlDscIsAgentOperator.ps1
@@ -30,23 +30,23 @@
 
     .EXAMPLE
         $serverObject = Connect-SqlDscDatabaseEngine -InstanceName 'MyInstance'
-        Test-SqlDscAgentOperator -ServerObject $serverObject -Name 'MyOperator'
+        Test-SqlDscIsAgentOperator -ServerObject $serverObject -Name 'MyOperator'
 
         Tests if the SQL Agent Operator named 'MyOperator' exists.
 
     .EXAMPLE
         $serverObject = Connect-SqlDscDatabaseEngine -InstanceName 'MyInstance'
-        $serverObject | Test-SqlDscAgentOperator -Name 'MyOperator' -EmailAddress 'admin@contoso.com'
+        $serverObject | Test-SqlDscIsAgentOperator -Name 'MyOperator' -EmailAddress 'admin@contoso.com'
 
         Tests if the SQL Agent Operator exists and has the specified email address using pipeline input.
 
     .EXAMPLE
         $serverObject = Connect-SqlDscDatabaseEngine -InstanceName 'MyInstance'
-        $serverObject | Test-SqlDscAgentOperator -Name 'MyOperator' -Refresh
+        $serverObject | Test-SqlDscIsAgentOperator -Name 'MyOperator' -Refresh
 
         Refreshes the server operators collection before testing if **MyOperator** exists.
 #>
-function Test-SqlDscAgentOperator
+function Test-SqlDscIsAgentOperator
 {
     [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseSyntacticallyCorrectExamples', '', Justification = 'Because the rule does not yet support parsing the code when a parameter type is not available. The ScriptAnalyzer rule UseSyntacticallyCorrectExamples will always error in the editor due to https://github.com/indented-automation/Indented.ScriptAnalyzerRules/issues/8.')]
     [CmdletBinding()]
@@ -67,20 +67,20 @@ function Test-SqlDscAgentOperator
         $Refresh
     )
 
-    # cSpell: ignore TSAO
+    # cSpell: ignore TISAO
     process
     {
-        Write-Verbose -Message ($script:localizedData.Test_SqlDscAgentOperator_TestingOperator -f $Name)
+        Write-Verbose -Message ($script:localizedData.Test_SqlDscIsAgentOperator_TestingOperator -f $Name)
 
         $operatorObject = Get-AgentOperatorObject -ServerObject $ServerObject -Name $Name -IgnoreNotFound -Refresh:$Refresh
 
         if (-not $operatorObject)
         {
-            Write-Verbose -Message ($script:localizedData.Test_SqlDscAgentOperator_OperatorNotFound -f $Name)
+            Write-Verbose -Message ($script:localizedData.Test_SqlDscIsAgentOperator_OperatorNotFound -f $Name)
             return $false
         }
 
-        Write-Verbose -Message ($script:localizedData.Test_SqlDscAgentOperator_OperatorFound -f $Name)
+        Write-Verbose -Message ($script:localizedData.Test_SqlDscIsAgentOperator_OperatorFound -f $Name)
 
         return $true
     }