+++ tests/Unit/Public/New-SqlDscAgentAlert.Tests.ps1
@@ -125,6 +125,7 @@ Describe 'New-SqlDscAgentAlert' -Tag 'Public' {
             $script:mockNewAlert.MessageID = 0
 
             # Mock the private functions
+            Mock -CommandName 'Test-SqlDscAgentAlert' -MockWith { return $true}
             Mock -CommandName 'Get-AgentAlertObject'
             Mock -CommandName 'Assert-BoundParameter'
         }